requirements.txt
Django==1.6.1
dj-database-url==0.2.2
dj-static==0.0.5
django-tastypie==0.11.0
django-toolbelt==0.0.1
gunicorn==18.0
psycopg2==2.5.2
python-dateutil==2.2
python-mimeparse==0.1.4
six==1.5.2
static==0.4
wsgiref==0.1.2
